feat(waitlist): add Waitlist section to the homepage
Added a new `Waitlist` component to the homepage for informing visitors about upcoming features and allowing them to join a waitlist. Included are new assets, 'Phone.png' and 'WaitlistVector.png', which are used in the new component. The `Waitlist` component has been imported into `index.vue` and added to the homepage template. This addition aims to engage visitors and build anticipation for future releases.